Move map-export models to model file
