+++ projects/map/src/lib/openlayers-map/openlayers-map.ts
@@ -3,7 +3,9 @@ import { Projection } from 'ol/proj';
 import { View } from 'ol';
 import { NgZone } from '@angular/core';
 import { defaults as defaultInteractions, DragPan, MouseWheelZoom } from 'ol/interaction';
-import { LayerManagerModel, MapViewDetailsModel, MapViewerModel, MapViewerOptionsModel, OlMapStyleType } from '../models';
+import {
+  LayerManagerModel, MapExportOptions, MapExportResult, MapViewDetailsModel, MapViewerModel, MapViewerOptionsModel, OlMapStyleType,
+} from '../models';
 import { ProjectionsHelper } from '../helpers/projections.helper';
 import { OpenlayersExtent } from '../models/extent.type';
 import { OpenLayersLayerManager } from './open-layers-layer-manager';
@@ -14,7 +16,6 @@ import { Size } from 'ol/size';
 import { ToolManagerModel } from '../models/tool-manager.model';
 import { OpenLayersToolManager } from './open-layers-tool-manager';
 import { OpenLayersEventManager } from './open-layers-event-manager';
-import { MapExportOptions, MapExportResult } from '../map-service/map.service';
 import { Feature } from 'ol';
 import { Geometry } from 'ol/geom';
 import { buffer, Extent, extend, getCenter } from 'ol/extent';
